+++ Cocoa/installation_scripts/unxv_h0licow.sh
@@ -62,35 +62,40 @@ if [ -z "${IGNORE_HOLICOW_STRONG_LENSING_DATA}" ]; then
 
   # ---------------------------------------------------------------------------
   # in case this script is called twice
+  if [ -n "${OVERWRITE_EXISTING_HOLICOW_DATA}" ]; then
   
-  rm -rf "${EDATAF:?}/${TMP:?}"
+    rm -rf "${EDATAF:?}/${TMP:?}"
 
-  rm -rf "${EDATAF:?}/${FOLDER:?}"
+    rm -rf "${EDATAF:?}/${FOLDER:?}"
   
-  # ---------------------------------------------------------------------------
+  fi
+  
+  if [ ! -d "${EDATAF:?}/${FOLDER:?}" ]; then
 
-  cdfolder "${EDATAF:?}" || return 1
+    cdfolder "${EDATAF:?}" || return 1
 
-  # check if the link exists
-  "${CURL:?}" -fsS "${URL:?}" \
-    >${OUT1:?} 2>${OUT2:?} || { error "${EC27:?} (URL=${URL:?})"; return 1; }
-    
-  ${GIT:?} clone "${URL:?}" "${TMP:?}" \
-    >${OUT1:?} 2>${OUT2:?} || { error "${EC15:?}"; return 1; }
+    # check if the link exists
+    "${CURL:?}" -fsS "${URL:?}" \
+      >${OUT1:?} 2>${OUT2:?} || { error "${EC27:?} (URL=${URL:?})"; return 1; }
+      
+    ${GIT:?} clone "${URL:?}" "${TMP:?}" \
+      >${OUT1:?} 2>${OUT2:?} || { error "${EC15:?}"; return 1; }
 
-  cdfolder "${EDATAF:?}/${TMP:?}" || return 1
+    cdfolder "${EDATAF:?}/${TMP:?}" || return 1
 
-  if [ -n "${HOLICOW_DATA_GIT_COMMIT}" ]; then
-    ${GIT:?} checkout "${HOLICOW_DATA_GIT_COMMIT:?}" \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC16:?}"; return 1; }
+    if [ -n "${HOLICOW_DATA_GIT_COMMIT}" ]; then
+      ${GIT:?} checkout "${HOLICOW_DATA_GIT_COMMIT:?}" \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC16:?}"; return 1; }
+    fi
+
+    mv "${FOLDER:?}" "${EDATAF:?}"
+
+    rm -rf "${EDATAF:?}/${TMP:?}"
+
   fi
 
-  mv "${FOLDER:?}" "${EDATAF:?}"
-
-  rm -rf "${EDATAF:?}/${TMP:?}"
-
-  # ---------------------------------------------------------------------------
-  
+  cdfolder "${ROOTDIR}" || return 1;
+    
   pbottom "SETUP/UNXV ${PRINTNAME:?} DATA" || return 1
 
   unset_all || return 1
