+++ Cocoa/installation_scripts/unxv_planck2018_basic.sh
@@ -54,31 +54,39 @@ if [ -z "${IGNORE_PLANCK_CMB_DATA}" ]; then
 
   # ---------------------------------------------------------------------------
   # note: in case this script is run twice
+  if [ -n "${OVERWRITE_EXISTING_PLANCK_CMB_DATA}" ]; then
+ 
+    rm -rf "${EDATAF:?}/planck/plc_3.0/low_l"
+    rm -rf "${EDATAF:?}/planck/plc_3.0/lensing"
+    rm -rf "${EDATAF:?}/planck/plc_3.0/hi_l/plik"
+    rm -rf "${EDATAF:?}/planck/plc_3.0/hi_l/plik_lite"
 
-  rm -rf "${EDATAF:?}/planck/plc_3.0/low_l"
-  rm -rf "${EDATAF:?}/planck/plc_3.0/lensing"
-  rm -rf "${EDATAF:?}/planck/plc_3.0/hi_l/plik"
-  rm -rf "${EDATAF:?}/planck/plc_3.0/hi_l/plik_lite"
+  fi
 
-  # ---------------------------------------------------------------------------
+  if [[ ! -d "${EDATAF:?}/planck/plc_3.0/low_l" && \
+        ! -d "${EDATAF:?}/planck/plc_3.0/lensing" && \
+        ! -d "${EDATAF:?}/planck/plc_3.0/hi_l/plik" && \
+        ! -d "${EDATAF:?}/planck/plc_3.0/hi_l/plik_lite" ]]; then
 
-  cdfolder "${EDATAF:?}/planck/plc_3.0" || return 1
-  
-  tar xf lensing.xz \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
-
-  tar xf low_l.xz \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
-
-  cdfolder "${EDATAF:?}/planck/plc_3.0/hi_l" || return 1
-  
-  tar xf plik.xz \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
-
-  tar xf plik_lite.xz \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+    cdfolder "${EDATAF:?}/planck/plc_3.0" || return 1
     
-  # ---------------------------------------------------------------------------
+    tar xf lensing.xz \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+
+    tar xf low_l.xz \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+
+    cdfolder "${EDATAF:?}/planck/plc_3.0/hi_l" || return 1
+    
+    tar xf plik.xz \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+
+    tar xf plik_lite.xz \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+    
+  fi
+
+  cdfolder "${ROOTDIR}" || return 1
 
   pbottom "SETUP/UNXV ${PRINTNAME:?} DATA" || return 1
