+++ Cocoa/installation_scripts/unxv_sn.sh
@@ -63,16 +63,21 @@ if [ -z "${IGNORE_SN_DATA}" ]; then
 
   # ---------------------------------------------------------------------------
   # note: in case script run >1x w/ previous run stoped prematurely b/c error
-
-  rm -rf "${PACKDIR:?}"
-
-  # ---------------------------------------------------------------------------
-
-  cdfolder "${EDATAF:?}" || return 1
-
-  tar xf "${FILE:?}" >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+  if [ -n "${OVERWRITE_EXISTING_SN_DATA}" ]; then
+    
+    rm -rf "${PACKDIR:?}"
   
-  # ---------------------------------------------------------------------------
+  fi
+
+  if [ ! -d "${PACKDIR:?}" ]; then
+  
+    cdfolder "${EDATAF:?}" || return 1
+
+    tar xf "${FILE:?}" >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?}"; return 1; }
+  
+  fi
+  
+  cdfolder "${ROOTDIR}" || return 1;
   
   pbottom "SETUP/UNXV ${PRINTNAME:?} DATA" || return 1
